Add TransactionCommand and source clean
